webapp/requirements.txt
Plotly==4.14.3
Streamlit==0.74.1
pandas==1.2.0
pillow==8.1.0